Use native file download instead of file-saver
